Change Route to be a top-level type alias
